refactor(mcp): share the automation-rule presenter and point missing ids at a list tool
